@@ -1,44 +1,45 @@
 {
   "name": "ui",
-  "version": "0.2.0",
+  "version": "next",
   "private": true,
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint",
-    "i18n:report": "vue-cli-service i18n:report --src \"./src/**/*.?(js|vue)\" --locales \"./src/locales/**/*.json\""
+    "dev": "vite",
+    "serve": "vite --open",
+    "preview": "vite preview",
+    "build": "vite build",
+    "lint": "eslint --ext .js,.vue --ignore-path .gitignore --fix src",
+    "lint:fix": "eslint src --fix",
+    "format": "prettier .  --write",
+    "i18n:report": "vite i18n:report --src \"./src/**/*.?(js|vue)\" --locales \"./src/locales/**/*.json\""
   },
   "dependencies": {
+    "@vueuse/core": "^6.1.0",
     "core-js": "^3.6.5",
-    "vue": "^2.6.11",
-    "vue-i18n": "^8.22.3",
-    "vue-router": "^3.2.0",
-    "vuetify": "^2.4.0",
-    "vuex": "^3.4.0"
+    "mitt": "^3.0.0",
+    "vue": "^3.2.20",
+    "vue-i18n": "^9.2.0-beta.12",
+    "vue-router": "^4.0.11",
+    "vuex": "^4.0.2"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-eslint": "~4.5.0",
-    "@vue/cli-plugin-router": "~4.5.0",
-    "@vue/cli-plugin-vuex": "^4.5.12",
-    "@vue/cli-service": "~4.5.0",
-    "@vue/eslint-config-prettier": "^6.0.0",
-    "babel-eslint": "^10.1.0",
-    "chart.js": "^2.9.4",
-    "eslint": "^6.7.2",
-    "eslint-plugin-prettier": "^3.3.1",
-    "eslint-plugin-vue": "^6.2.2",
+    "@headlessui/vue": "^1.4.1",
+    "@heroicons/vue": "^1.0.4",
+    "@intlify/vite-plugin-vue-i18n": "^2.4.0",
+    "@j-t-mcc/vue3-chartjs": "^1.1.4",
+    "@rollup/plugin-alias": "^3.1.5",
+    "@vitejs/plugin-vue": "^1.9.3",
+    "chart.js": "^3.5.0",
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-vue": "^7.19.1",
+    "eslint-plugin-html": "^6.2.0",
     "lodash-es": "^4.17.21",
-    "node-sass": "^4.12.0",
-    "prettier": "^2.2.1",
-    "sass": "^1.32.0",
-    "sass-loader": "^10.0.0",
-    "socket.io-client": "^4.0.1",
-    "vue-chartjs": "^3.5.1",
-    "vue-cli-plugin-i18n": "~2.0.3",
-    "vue-cli-plugin-vuetify": "~2.3.1",
-    "vue-template-compiler": "^2.6.11",
-    "vuetify-loader": "^1.7.0",
-    "webpack-remove-debug": "^0.1.0"
+    "prettier": "^2.3.2",
+    "sass": "^1.42.1",
+    "socket.io-client": "^4.1.3",
+    "vite": "^2.6.7",
+    "vite-plugin-windicss": "^1.3.0",
+    "windicss": "^3.1.9"
   }
 }
