@@ -1,96 +1,96 @@
 <template>
-  <v-card class="fill-height">
-    <v-card-title>{{ $t("rooms.title") }}</v-card-title>
-    <v-data-table :headers="headers" :items="roomsAsObjects" dense>
-      <template v-slot:item.actions="{ item }">
-        <v-tooltip bottom v-if="isSocketLeaveSupported">
-          <template v-slot:activator="{ on, attrs }">
-            <v-btn
-              v-bind="attrs"
-              v-on="on"
-              @click="leave(item)"
+  <div class="p-4 bg-white shadow-lg rounded-lg my-20 mx-6">
+    <div class="font-bold text-left p-4">{{ t("rooms.title") }}</div>
+    <Datatable :headers="headers" :items="roomsAsObjects" :dense="true">
+      <template #actions="{ item }">
+        <tooltip v-if="isSocketLeaveSupported">
+          <template #activator>
+            <button
+              class="btn ml-3"
               :disabled="isReadonly"
-              small
-              class="ml-3"
+              @click="leave(item)"
             >
-              <v-icon>mdi-tag-off-outline</v-icon>
-            </v-btn>
+              <LogoutIcon />
+            </button>
           </template>
-          <span>{{ $t("sockets.leave") }}</span>
-        </v-tooltip>
+          <span>{{ t("sockets.leave") }}</span>
+        </tooltip>
 
-        <v-tooltip bottom>
-          <template v-slot:activator="{ on, attrs }">
-            <v-btn
-              v-bind="attrs"
-              v-on="on"
+        <tooltip>
+          <template #activator>
+            <button
+              class="btn ml-3"
               @click="displayDetails(item)"
-              small
-              class="ml-3"
             >
-              <v-icon>mdi-dots-horizontal</v-icon>
-            </v-btn>
+              <DotsHorizontalIcon />
+            </button>
           </template>
-          <span>{{ $t("rooms.displayDetails") }}</span>
-        </v-tooltip>
+          <span>{{ t("rooms.displayDetails") }}</span>
+        </tooltip>
       </template>
-    </v-data-table>
-
-    <v-card-text>
+    </Datatable>
+    <div class="p-4">
       <form @submit.prevent="onSubmit">
-        <v-combobox
-          :search-input.sync="newRoom"
-          :label="$t('sockets.join-a-room')"
+        <CustomCombobox
+          v-model="newRoom"
+          class="select-room"
+          :label="t('sockets.join-a-room')"
           :items="availableRooms"
-          item-value="name"
-          item-text="name"
-          class="select-room d-inline-block mr-3"
           :disabled="isReadonly"
-          :return-object="false"
         />
 
-        <v-tooltip bottom>
-          <template v-slot:activator="{ on, attrs }">
-            <v-btn
-              v-bind="attrs"
-              v-on="on"
+        <tooltip>
+          <template #activator>
+            <button
+              class="btn"
               type="submit"
-              small
               :disabled="isReadonly"
             >
-              <v-icon>mdi-tag-plus-outline</v-icon>
-            </v-btn>
+              <PlusCircleIcon />
+            </button>
           </template>
-          <span>{{ $t("sockets.join") }}</span>
-        </v-tooltip>
+          <span>{{ t("sockets.join") }}</span>
+        </tooltip>
       </form>
-    </v-card-text>
-  </v-card>
+    </div>
+  </div>
 </template>
 
 <script>
-import { mapGetters, mapState } from "vuex";
-import SocketHolder from "../../SocketHolder";
+import { ref, computed } from "vue";
+import { useStore } from "vuex";
+import { useI18n } from "vue-i18n";
+import { useRouter } from "vue-router";
 import { differenceBy } from "lodash-es";
+import { DotsHorizontalIcon } from "@heroicons/vue/solid";
+import { LogoutIcon, PlusCircleIcon } from "@heroicons/vue/outline";
+
+import SocketHolder from "@/SocketHolder";
+import Tooltip from "@/components/Tooltip.vue";
+import Datatable from "@/components/Datatable.vue";
+import CustomCombobox from "@/components/CustomCombobox.vue";
 
 export default {
   name: "SocketRooms",
-
+  components: {
+    Tooltip,
+    Datatable,
+    CustomCombobox,
+    PlusCircleIcon,
+    LogoutIcon,
+    DotsHorizontalIcon,
+  },
   props: {
     socket: Object,
   },
-
-  data() {
-    return {
-      newRoom: "",
-    };
-  },
-
-  computed: {
-    headers() {
-      return [
+  setup(props) {
+    const { t } = useI18n();
+    const store = useStore();
+    const router = useRouter();
+    const newRoom = ref("");
+    const headers = computed(() => [
         {
-          text: this.$t("id"),
+          text: t("id"),
           value: "name",
           align: "start",
         },
@@ -99,69 +99,69 @@ export default {
           align: "end",
           sortable: false,
         },
-      ];
-    },
-    roomsAsObjects() {
-      return this.socket.rooms
+      ]);
+    const roomsAsObjects = computed(() => props.socket.rooms
         .slice(0)
         .sort()
         .map((room) => ({
           name: room,
-        }));
-    },
-    availableRooms() {
-      return differenceBy(
-        this.findRoomsByNamespace(this.socket.nsp),
-        this.roomsAsObjects,
+        })));
+    const availableRooms = computed(() => differenceBy(
+        store.getters['main/findRoomsByNamespace'](props.socket.nsp),
+        roomsAsObjects.value,
         "name"
-      );
-    },
-    ...mapState({
-      selectedNamespace: (state) => state.main.selectedNamespace,
-      isReadonly: (state) => state.config.readonly,
-      isSocketLeaveSupported: (state) =>
-        state.config.supportedFeatures.includes("LEAVE"),
-      isSocketDisconnectSupported: (state) =>
-        state.config.supportedFeatures.includes("DISCONNECT"),
-    }),
-    ...mapGetters("main", ["findRoomsByNamespace"]),
-  },
+      ));
+    const selectedNamespace = computed(() => store.state.main.selectedNamespace);
+    const isReadonly = computed(() => store.state.config.readonly);
+    const isSocketLeaveSupported = computed(() => store.state.config.supportedFeatures.includes("LEAVE"));
+    const isSocketDisconnectSupported = computed(() => store.state.config.supportedFeatures.includes("DISCONNECT"));
+    // const findRoomsByNamespace = computed(() => store.getters['main/findRoomsByNamespace']);
 
-  methods: {
-    emit() {
+    const emit = () => {
       // TODO
-    },
-    onSubmit() {
+    };
+    const onSubmit = () => {
       SocketHolder.socket.emit(
         "join",
-        this.socket.nsp,
-        this.newRoom,
-        this.socket.id
+        props.socket.nsp,
+        newRoom.value,
+        props.socket.id
       );
-      this.newRoom = "";
-    },
-    leave(room) {
-      SocketHolder.socket.emit(
+      newRoom.value = "";
+    };
+    const leave = (room) => SocketHolder.socket.emit(
         "leave",
-        this.socket.nsp,
+        props.socket.nsp,
         room.name,
-        this.socket.id
+        props.socket.id
       );
-    },
-    disconnect() {
-      SocketHolder.socket.emit(
+    const disconnect = () => SocketHolder.socket.emit(
         "_disconnect",
-        this.socket.nsp,
+        props.socket.nsp,
         false,
-        this.socket.id
+        props.socket.id
       );
-    },
-    displayDetails(room) {
-      this.$router.push({
+    const displayDetails = (room) => router.push({
         name: "room",
-        params: { nsp: this.socket.nsp, name: room.name },
+        params: { nsp: props.socket.nsp, name: room.name },
       });
-    },
+
+    return {
+      t,
+      newRoom,
+      headers,
+      roomsAsObjects,
+      availableRooms,
+      selectedNamespace,
+      isReadonly,
+      isSocketLeaveSupported,
+      isSocketDisconnectSupported,
+      emit,
+      onSubmit,
+      leave,
+      disconnect,
+      displayDetails,
+    };
   },
 };
 </script>
