@@ -1,13 +1,15 @@
 module.exports = {
-  root: true,
   env: {
     node: true,
   },
-  extends: ["plugin:vue/essential", "eslint:recommended", "@vue/prettier"],
-  parserOptions: {
-    parser: "babel-eslint",
-  },
+  extends: [
+    'eslint:recommended',
+    'plugin:vue/vue3-recommended',
+    '**prettier**',
+  ],
   rules: {
+    "no-var": "error",
+    "no-unused-vars": "warn",
     "no-console": process.env.NODE_ENV === "production" ? "warn" : "off",
     "no-debugger": process.env.NODE_ENV === "production" ? "warn" : "off",
   },
