@@ -1,15 +1,34 @@
+<template>
+  <div>
+    <vue3-chart-js
+      type="doughnut"
+      :data="chartData"
+      :options="options"
+    ></vue3-chart-js>
+  </div>
+</template>
+
 <script>
-import { Doughnut, mixins } from "vue-chartjs";
-const { reactiveProp } = mixins;
+import Vue3ChartJs from "@j-t-mcc/vue3-chartjs";
 
 export default {
-  extends: Doughnut,
-  mixins: [reactiveProp],
-  props: ["options"],
-  mounted() {
-    this.renderChart(this.chartData, {
-      legend: false,
-    });
+  name: "Doughnut",
+  components: {
+    Vue3ChartJs,
+  },
+  props: {
+    chartData: {
+      type: Object,
+      required: true,
+    },
+  },
+  setup() {
+    const options = {
+      plugins: {
+        legend: false,
+      }
+    };
+    return { options };
   },
 };
 </script>
