@@ -1,21 +1,26 @@
 <template>
   <Status
     :value="healthy"
-    :ok-label="$t('servers.healthy')"
-    :ko-label="$t('servers.unhealthy')"
+    :ok-label="t('servers.healthy')"
+    :ko-label="t('servers.unhealthy')"
   />
 </template>
 
 <script>
-import Status from "./Status";
+import { useI18n } from "vue-i18n";
+
+import Status from "@/components/Status.vue";
 
 export default {
   name: "ServerStatus",
-
-  components: { Status },
-
+  components: { Status, },
   props: {
     healthy: Boolean,
+  },
+  setup() {
+    const { t } = useI18n();
+
+    return { t };
   },
 };
 </script>
