@@ -1,38 +1,54 @@
 <template>
-  <v-chip small :color="color" outlined>{{ label }}</v-chip>
+  <div
+    class="
+      rounded-full
+      border-1
+      px-2
+      mx-1.5
+      text-sm
+      cursor-default
+      hover:(bg-gray-400/10)
+    "
+    :class="color"
+  >
+    {{ label }}
+  </div>
 </template>
 
 <script>
-import colors from "vuetify/lib/util/colors";
+import { computed } from "vue";
 
 export default {
   name: "Transport",
-
   props: {
-    transport: String,
+    transport: {
+      type: String,
+      required: true,
+    },
   },
-
-  computed: {
-    label() {
-      switch (this.transport) {
+  setup(props) {
+    const label = computed(() => {
+      switch (props.transport) {
         case "polling":
           return "HTTP long-polling";
         case "websocket":
           return "WebSocket";
         default:
-          return this.transport;
+          return props.transport;
       }
-    },
-    color() {
-      switch (this.transport) {
+    });
+
+    const color = computed(() => {
+      switch (props.transport) {
         case "polling":
-          return colors.orange.base;
+          return "border-yellow-600 text-yellow-600";
         case "websocket":
-          return colors.green.base;
+          return "border-green-500 text-green-500";
         default:
-          return colors.grey.base;
+          return "border-gray-400 text-gray-400";
       }
-    },
+    });
+    return { label, color };
   },
 };
 </script>
